mise en place des flash message

+++ app/Http/Controllers/UsersController.php
@@ -19,7 +19,7 @@ class UsersController extends Controller
         return view('pages.dashboard.account', ["userDetails" => $userDetails, "planDetails" => $planDetails]);
     }
 
-    function updateUser(Request $request)
+    function updateUser(Request $request, $id)
     {
         $request->validate([
             'firstname' => 'required',
@@ -28,18 +28,25 @@ class UsersController extends Controller
             'password' => 'required',
         ]);
 
-        if ($request->user()->id === $request->id || $request->user()->is_admin) {
-            $user = User::where('id', $request->user()->id);
+        if ($request->user()->id === $id || $request->user()->is_admin) {
+            $user = User::where('id', $id);
 
+            if ($request->user()->is_admin && $request->is_admin) {
+                $is_admin = $request->is_admin == "on" ? 1 : 0;
+            } else {
+                $is_admin = 0;
+            }
+            // dd($request->all());
             $user->update([
                 'firstname' => $request->firstname,
                 'lastname' => $request->lastname,
                 'email' => $request->email,
-                'password' => Hash::make($request->password)
+                'password' => Hash::make($request->password),
+                'is_admin' => $is_admin
             ]);
         }
         if ($request->user()->is_admin) {
-            return redirect()->route('admin.users');
+            return redirect()->route('admin.users')->with('success', 'Utilisateur modifié avec succès');
         } else {
             return redirect()->route('account');
         }
@@ -59,10 +66,10 @@ class UsersController extends Controller
     }
 
 
-    function renderUserAdminEditForm(Request $request)
+    function renderUserAdminEditForm(Request $request, $id)
     {
         if ($request->user()->is_admin) {
-            $userInfo = User::firstWhere('id', $request->user()->id);
+            $userInfo = User::firstWhere('id', $id);
             return view('pages.dashboard.settings', ['userInfo' => $userInfo]);
         } else {
             return redirect()->route('admin.users');
@@ -71,17 +78,31 @@ class UsersController extends Controller
 
     function deleteUser(Request $request)
     {
-        $deletedArticle = User::firstWhere('id', $request->id);
-        $deletedArticle->delete();
-        return redirect()->route('admin.users');
+        if ($request->user()->is_admin) {
+            $deletedArticle = User::firstWhere('id', $request->id);
+            if (isset($deletedArticle)) {
+                $deletedArticle->delete();
+                return redirect()->route('admin.users')->with('success', 'Utilisateur supprimé avec succès');
+            }
+            return redirect()->route('admin.users')->with('error', 'Utilisateur introuvable.');
+        } else {
+            return redirect()->route('acccount');
+        }
     }
 
 
-    function renderUserDetails(Request $request)
+    function renderUserDetails(Request $request, $id)
     {
+
         if ($request->user()->is_admin) {
-            $userDetails = User::firstWhere('id', $request->user()->id);
-            return view('pages.dashboard.account', ["userDetails" => $userDetails]);
+            $userDetails = User::firstWhere('id', $id);
+            if (isset($userDetails->subscriptions[0]->stripe_plan)) {
+                $planDetails = Plan::where('stripe_id', $userDetails->subscriptions[0]->stripe_plan)->first();
+            } else {
+                $planDetails = [];
+            }
+
+            return view('pages.dashboard.account', ["userDetails" => $userDetails, "planDetails" => $planDetails]);
         } else {
             return redirect()->route('admin.users');
         }
